Move endpoint scanner into the policy API
