+++ samples/Authorization.Sample/Controllers/DocsController.cs
@@ -12,51 +12,9 @@ namespace Authorization.Sample.Controllers
         [Route("docs")]
         public IHttpActionResult Get()
         {
-            var explorer = new ApiExplorer(Configuration);
-
-            var endpoints = new List<EndpointSummary>();
-            foreach (var description in explorer.ApiDescriptions)
-            {
-                var action = description.ActionDescriptor;
-
-                var endpoint = new EndpointSummary
-                {
-                    HttpMethod = description.HttpMethod.Method,
-                    Controller = description.ActionDescriptor.ControllerDescriptor.ControllerName,
-                    Action = description.ActionDescriptor.ActionName,
-                    EndpointPath = $"/{description.RelativePath}"
-                };
-
-
-                var pipeline = action.GetFilterPipeline();
-
-                foreach (var filter in pipeline.OfType<PolicyAttribute>())
-                {
-                    endpoint.Policies.Add(filter.PolicyType.Name);
-                }
-
-                endpoints.Add(endpoint);
-            }
+            var endpoints = Configuration.ScanEndpointsForPolicyRequirements();
 
             return Json(endpoints);
-        }
-
-        private class EndpointSummary
-        {
-            public EndpointSummary()
-            {
-                Policies = new HashSet<string>();
-            }
-
-            public string HttpMethod { get; set; }
-
-            public string Controller { get; set; }
-
-            public string Action { get; set; }
-
-            public string EndpointPath { get; set; }
-
-            public ICollection<string> Policies { get; }
         }
     }
 }
