+++ src/Authorization.WebApi/Authorization.WebApi.csproj
@@ -56,6 +56,7 @@
   <ItemGroup>
     <Compile Include="AuthenticateAttribute.cs" />
     <Compile Include="ClaimPolicyAttribute.cs" />
+    <Compile Include="HttpConfigurationExtensions.cs" />
     <Compile Include="InRolePolicyAttribute.cs" />
     <Compile Include="PolicyAttribute.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
